GH-56 Fix #56 issue of TestNG test suit not running in IDE due to BaseTest being in inaccessible package
